+++ bindings/python/CHANGELOG.rst
@@ -4,8 +4,9 @@ Changelog
 Changes in Version 1.6.0
 ------------------------
 
-- Bundle libmongocrypt 1.8.0-alpha1 in release wheels.
-- **Remove support for libmongocrypt <=1.8.0, libmongocrypt >=1.8.0-alpha1
+- Drop support for Python 2 and Python <3.7. Python >=3.7 is now required.
+- Bundle libmongocrypt 1.8.0 in release wheels.
+- **Remove support for libmongocrypt <=1.8.0, libmongocrypt >=1.8.0
   is now required.** Note this is only relevant for users that install from
   source or use the ``PYMONGOCRYPT_LIB`` environment variable.